12.0.0.20230529: refactor librarian
