+++ librarian/CMakeLists.txt
@@ -1,10 +1,7 @@
 set(libs sysy;sysu)
 foreach(lib IN LISTS libs)
-  file(GLOB mains ${lib}/*.cc)
+  file(GLOB mains lib/${lib}/*.cc)
   add_library(${lib} SHARED ${mains})
   install(TARGETS ${lib})
-  file(GLOB mains ${lib}/*.h)
-  foreach(mainfile ${mains})
-    install(FILES ${mainfile} DESTINATION include/${lib})
-  endforeach()
 endforeach()
+install(FILES include DESTINATION include)
