+++ librarian/lib/sysu/stdio.cc
@@ -1,4 +1,4 @@
-#include "stdio.h"
+#include "sysu/stdio.h"
 #include <stdio.h>
 
 #ifdef __cplusplus
