+++ librarian/lib/sysu/stdlib.cc
@@ -1,4 +1,4 @@
-#include "stdlib.h"
+#include "sysu/stdlib.h"
 #include <stdlib.h>
 
 #ifdef __cplusplus
