+++ librarian/lib/sysu/string.cc
@@ -1,4 +1,4 @@
-#include "string.h"
+#include "sysu/string.h"
 #include <string.h>
 
 #ifdef __cplusplus
