+++ tester/function_test2020/CMakeLists.txt
@@ -5,6 +5,6 @@ foreach(mainfile ${mains})
   add_executable(sysu-tester-function_test2020_${mainname} ${mainfile})
   target_link_libraries(sysu-tester-function_test2020_${mainname} sysy)
   target_include_directories(sysu-tester-function_test2020_${mainname}
-                             PRIVATE ${CMAKE_SOURCE_DIR}/librarian)
+                             PRIVATE ${CMAKE_SOURCE_DIR}/librarian/include)
   install(TARGETS sysu-tester-function_test2020_${mainname})
 endforeach()
