+++ tester/h_functional/CMakeLists.txt
@@ -5,6 +5,6 @@ foreach(mainfile ${mains})
   add_executable(sysu-tester-h_functional_${mainname} ${mainfile})
   target_link_libraries(sysu-tester-h_functional_${mainname} sysy)
   target_include_directories(sysu-tester-h_functional_${mainname}
-                             PRIVATE ${CMAKE_SOURCE_DIR}/librarian)
+                             PRIVATE ${CMAKE_SOURCE_DIR}/librarian/include)
   install(TARGETS sysu-tester-h_functional_${mainname})
 endforeach()
